texk/dvipdfm-x: Fix a bug in PK font support (S. Hirata)
git-svn-id: svn://tug.org/texlive/trunk/Build/source@35461 c570f23f-e606-0410-a88d-b1316a301751

+++ texk/dvipdfm-x/pkfont.c
@@ -613,8 +613,8 @@ pdf_font_load_pkfont (pdf_font *font)
 
         /* Update font BBox info */
         bbox.llx = MIN(bbox.llx, -pkh.bm_hoff);
-        bbox.lly = MIN(bbox.lly,  pkh.bm_voff - pkh.bm_ht);
-        bbox.urx = MAX(bbox.urx,  pkh.bm_wd - pkh.bm_hoff);
+        bbox.lly = MIN(bbox.lly,  (double)pkh.bm_voff - (double)pkh.bm_ht);
+        bbox.urx = MAX(bbox.urx,  (double)pkh.bm_wd - (double)pkh.bm_hoff);
         bbox.ury = MAX(bbox.ury,  pkh.bm_voff);
 
         pkt_ptr = NEW(pkh.pkt_len, unsigned char);
